refactor: Side 패키지 변경
