+++ src/main/java/janggi/domain/piece/Side.java
@@ -1,4 +1,4 @@
-package janggi.domain;
+package janggi.domain.piece;
 
 public enum Side {
     HAN, CHO
