+++ src/main/java/janggi/domain/piece/movement/MovementStrategy.java
@@ -1,8 +1,8 @@
 package janggi.domain.piece.movement;
 
-import janggi.domain.Side;
 import janggi.domain.piece.Pieces;
 import janggi.domain.piece.Position;
+import janggi.domain.piece.Side;
 
 public interface MovementStrategy {
