+++ src/main/java/janggi/domain/piece/movement/dynamic/PawnMovementStrategy.java
@@ -1,8 +1,8 @@
 package janggi.domain.piece.movement.dynamic;
 
-import janggi.domain.Side;
 import janggi.domain.piece.Pieces;
 import janggi.domain.piece.Position;
+import janggi.domain.piece.Side;
 import java.util.Set;
 
 public class PawnMovementStrategy implements DynamicMovementStrategy {
