+++ src/main/java/janggi/domain/piece/movement/fixed/FixedMovementStrategy.java
@@ -1,8 +1,8 @@
 package janggi.domain.piece.movement.fixed;
 
-import janggi.domain.Side;
 import janggi.domain.piece.Pieces;
 import janggi.domain.piece.Position;
+import janggi.domain.piece.Side;
 import janggi.domain.piece.movement.MovementStrategy;
 
 public interface FixedMovementStrategy extends MovementStrategy {
