+++ src/test/java/janggi/domain/piece/movement/dynamic/PawnMovementStrategyTest.java
@@ -3,8 +3,8 @@ package janggi.domain.piece.movement.dynamic;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import janggi.domain.ReplaceUnderBar;
-import janggi.domain.Side;
 import janggi.domain.piece.Position;
+import janggi.domain.piece.Side;
 import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
